Deletes game objects that leave the screen
The shooter will be unable to leave the screen because the move method
will detect if the shooter has moved past the edge of the screen or not,
and if so, it will be repositioned.

The lasers will simply be allowed to fly off the screen.

The game loop will detect if any GameObject has left the screen; if
detected they will be deleted.