Changes how sprites are updated over time
Rather than call a move function for every sprite in the game a
function called update is now called.

The key press logic for the arrow keys was moved into the
shooter sprite object because nobody else needs it.

The new update logic makes it easier to detect when a sprite
has left the screen so we don't need a redundant withinScreen(...)
for every sprite anymore - and it will be easier to delete any
sprites in the future because you can call kill() in update if you
have to.